Remove our customized shibboleth plugin
Gluu has merged this component into the oxShibboleth
project so the upstream repo is no longer maintained.

THe original purpose of our mod was to support affilliations
of SAML SPs, but with the new idp interception script feature
there is probably a better way to manage this that does not
require customization.